feat(Core): Move the ProgressBar to helpers
