+++ package.json_off
@@ -1,6 +1,6 @@
 {
   "name": "mediafiredl",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "description": "A simple command line software to download mediafire files",
   "main": "index.js",
   "type": "module",
@@ -36,7 +36,6 @@
     "utf8": "^3.0.0",
     "commander": "^12.1.0",
     "i18n": "^0.15.1",
-    "os-locale": "^6.0.2",
     "chalk": "^5.3.0",
     "cli-progress": "^3.12.0"
   },
