+++ src/helpers/ProgressBar/ProgressBar.test.ts
