+++ src/utils/getOsLanguage/getOsLanguage.ts
@@ -1,4 +1,4 @@
 export const getOsLanguage = (): string => {
-  const locale = Intl.DateTimeFormat().resolvedOptions().locale;
-  return locale.split('-')[0];
-}
+	const locale = Intl.DateTimeFormat().resolvedOptions().locale;
+	return locale.split('-')[0];
+};
